requirements.txt
langchain
langgraph
pydantic
python-dotenv
requests
pytest
loguru
mistralai
langchain-mistralai